оптимизировал сортировку
