+++ project/src/components/header/header.tsx
@@ -5,8 +5,9 @@ import {AppRoute} from '../../const/const';
 import { logoutAction } from '../../store/api-action';
 import Logo from '../logo/logo';
 import { getAuthorizationStatus } from '../../store/user-process/selectors';
+import { memo } from 'react';
 
-export default function Header(): JSX.Element {
+function Header(): JSX.Element {
   const dispatch = useAppDispatch();
   const authorizationStatus = useAppSelector(getAuthorizationStatus);
   return (
@@ -30,7 +31,7 @@ export default function Header(): JSX.Element {
                     evt.preventDefault();
                     dispatch(logoutAction());
                   }}
-                  to={AppRoute.Main}
+                  to={AppRoute.Login}
                   >
                     <span className="header__signout">Sign out</span>
                   </Link>
@@ -49,3 +50,5 @@ export default function Header(): JSX.Element {
     </header>
   );
 }
+
+export default memo(Header);
