+++ project/src/components/logo/logo.tsx
@@ -1,3 +1,4 @@
+import { memo } from 'react';
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../const/const';
 
@@ -8,11 +9,11 @@ type LogoProps = {
 function Logo({classAction}: LogoProps): JSX.Element {
   return (
     <div className="header__left">
-      <Link className={classAction ? `header__logo-link ${classAction}` : 'header__logo-link'} to={AppRoute.Main}>
+      <Link className={classAction ? `header__logo-link ${classAction}` : 'header__logo-link'} to={AppRoute.Login}>
         <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
       </Link>
     </div>
   );
 }
-export default Logo;
+export default memo(Logo);
